Huggingface API with Unity
Installing and integrating the huggingface API for Unity. Got the automatic speech recognition (ASR) example to work with the whisper-tiny model. Needed to modify the example script to make it work. Worth noting this and also the fact some of the url endpoints for the API in the huggingface window may not work out the box.