Voice Recognition Code Optimization
Finished optimizing the code for handling the voice recogniton with the ASR model. Was able to implement a circular buffer to handle incoming audio samples instead of dynamically allocating an array each frame update. With the optimization and WebRtc VAD, everything is working well and the ASR model has proper data to recognize speech. Continuous voice recognition is now implemented with this.