Implementing Sentence Similarity Model 1.2
Got the Sentence Similarity model working through the API and integrated with the current Voice Interaction System. Updated the system to a decent starting point to make voice interaction more robust and abstract with the Sentence Similarity model. Implemented a Command Definition and Command Manager system where commands are general string inputs that are intended to invoke a function/action. The Sentence Similarity model compares recognized voice input to the collection of commands to find the most similar one above a specified threshold, currently 0.6.