Changed the VAD Algorithm and Added 2 More LoS
Changed the VAD algorithm from the WebRTCVAD plugin I was using before to a simple energy threshold root mean squared calculation. With this, there shouldn't be any more build issues for the headset since the old plugin was incompatible with Android. Also added 2 more levels of scale to the model. The first is a vein on the heart's exterior and the second is blood vessels within that vein.