Application build stuttering on startup (Quest 3)
Working to fix stuttering when the application starts in the quest 3.
Not sure why it happens specifically. Testing in play mode on the PC there is no stuttering or issue when the editor starts runtime.
Currently thinking it has to do with the shader to clip things outside the bounding box.
Created a shader variant collection and added it to preloaded shaders to try to fix.
Seems to help very slightly, but the application still stutters for a bit.